added: inc\current_process_modal_registry.h ...
...
modified:
* inc\game_actor_entity.h
* src\main.c
