+++ src/main.c
@@ -13,121 +13,8 @@
 #include <string.h>
 #endif
 
-#define WINDOW_SCREEN_SIZE_WIDTH 600
-#define WINDOW_SCREEN_SIZE_HEIGHT 480
-#define WINDOW_APPLICATION_TITLE "High Hunt!"
-// FPS is defined as so to be base-10:
-#define FPS_SET_TARGET 60
-// FPS target is 10, but -2 for 8 to reflect spritesheet animation indices
-// +2 additional frames for collision and invincibility
-
-enum CurrentProcessorModes {
-  USER_MODE = 0b10000,
-  FIQ_MODE = 0b10001,
-  IRQ_MODE = 0b10010,
-  SUPERVISOR_MODE = 0b10011,
-  ABORT_MODE = 0b10111,
-  UNDEFINED_MODE = 0b11011,
-  SYSTEM_MODE = 0b11111,
-};
-// the other 20 remaining registers
-enum ArmRegisters {
-  R0,
-  R1,
-  R2,
-  R3,
-  R4,
-  R5,
-  R6,
-  R7,
-  R8,
-  R9,
-  R10,
-  R11,
-  R12,
-  R13,
-  R14,
-  R15,
-  REG_IP = 12,
-  REG_SP = 13,
-  REG_LR = 14,
-  REG_PC = 15,
-}; // 32 registers in total ends here
-
-enum ArmConditions {
-  COND_EQ,
-  COND_NE,
-  COND_CS,
-  COND_CC,
-  COND_MI,
-  COND_PL,
-  COND_VS,
-  COND_VC,
-  COND_HI,
-  COND_LS,
-  COND_GE,
-  COND_LT,
-  COND_GT,
-  COND_LE,
-  COND_AL,
-};
-
-enum KnownInstructions {
-  INST_ADD_IMMEDIATE,
-  INST_B_ADDRESS,
-  INST_BL_ADDRESS,
-  INST_BLX_ADDRESS,
-  INST_BX_REGISTER,
-  INST_MOV_REGISTER,
-  INST_MOVT_IMMEDIATE,
-  INST_MOVW_IMMEDIATE,
-  INST_MVN_IMMEDIATE,
-  INST_POP_REGMASK,
-  INST_PUSH_REGMASH,
-  INST_SUB_IMMEDIATE,
-  INST_SVC_IMMEDIAT,
-  N_KNOWN_INSTRUCTIONS,
-};
-
-enum ArgumentTypes {
-  ARG_INVALID,
-  ARG_CONDITION,
-  ARG_REGISTER,
-  ARG_IMMEDIATE,
-  ARG_ADDRESS,
-  ARG_DATA_SYMBOL_ADDRESS,
-  ARG_DATA_SYMBOL_ADDRESS_TOP16,    // half of the 32 ...
-  ARG_DATA_SYMBOL_ADDRESS_BOTTOM16, // ... other half of the 32 registers
-  ARG_DATA_SYMBOL_SIZE,
-  ARG_FRAME_ADDRESS,
-  ARG_FRAME_ADDRESS_PC_RELATIVE,
-  ARG_REGMASK,
-};
-
-typedef struct Collision {
-  Rectangle hitbox;
-} Collision;
-
-typedef struct Position {
-  Vector2 position;
-} Position;
-
-typedef struct Health {
-  uint8_t health;
-} Health;
-
-typedef struct PlayerServiceLocator {
-  void *destination, *source, *data;
-  uint16_t size;
-} PlayerServiceLocator; // TODO: turn into a temporary address array system
-
-typedef struct PlayerServiceAvails {
-  PlayerServiceLocator *live_services;
-} PlayerServiceAvails;
-
-typedef struct AnnulPlayerService {
-  PlayerServiceLocator *player_service_locator;
-} AnnulPlayerService;
+#include "../inc/current_process_modal_registry.h"
+#include "../inc/game_actor_entity.h"
 
 int annul_player_service_location(
     PlayerServiceLocator *player_service_locator) {
@@ -135,46 +22,12 @@ int annul_player_service_location(
   player_service_locator = NULL;
   return 0;
 }
+// TODO(Daniel): utilise these codes from the included header
+// void execute_command(InputCommand command); 
+// void handle_command_inputs(InputHandler handler);
+// void execute_game_actor(InputCommand *command, PlayerEntity *player);
 
-typedef struct PlayerEntity {
-  PlayerServiceLocator *player_service_locator;
-} PlayerEntity;
-
-typedef struct InputCommand {
-  void *(command);
-} InputCommand;
-
-void execute_command(InputCommand command);
-
-typedef struct AttackCommand {
-  InputCommand *command;
-} AttackCommand;
-
-typedef struct DelayCommand {
-  InputCommand *command;
-} DelayCommand;
-
-typedef struct SustainCommand {
-  InputCommand *command;
-} SustainCommand;
-
-typedef struct ReleaseCommand {
-  InputCommand *command;
-} ReleaseCommand;
-
-typedef struct InputHandler {
-  InputCommand *attack_command;
-  InputCommand *delay_command;
-  InputCommand
-      *sustain_command; // call the actual root member which manipulates
-  InputCommand *release_command; // don't use the actual structs to call
-} InputHandler;
-
-void handle_command_inputs(InputHandler handler);
-
-void execute_game_actor(InputCommand *command, PlayerEntity *player);
-
-int initd_checked_player_instance(PlayerServiceLocator *service_instance) {
+int check_init_player_instance(PlayerServiceLocator *service_instance) {
   struct PlayerEntity {
     PlayerServiceLocator player_service_locator;
   } player_entity = {
@@ -200,21 +53,6 @@ int annul_player_entity_instance(PlayerEntity *player_entity) {
   player_entity = NULL;
   return 0;
 }
-
-typedef struct PlayerName {
-  const char *name;
-} PlayerName;
-
-typedef struct PlayerBundle {
-  PlayerEntity entity;
-  Position position;
-  Health health;
-  Texture2D sprite;
-} PlayerBundle;
-
-typedef struct PlayerObserver {
-  void(*notifier);
-} PlayerObserver;
 
 int (*on_notice)(const PlayerEntity entity, enum ArmRegisters registry);
